Correct code for advance task 1

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,7 +8,7 @@
  */
 int main(void)
 {
-	unsigned long long i;
+	unsigned long i;
 	int n = 2;
 
 	i = 612852475143;
